requirements.txt
Cython==0.29.21
matplotlib==3.3.1
tensorboard==2.0.0
kiwipiepy==0.8.1
librosa==0.8.0
numpy==1.20.3
scipy==1.6.3
Unidecode==1.1.1
omegaconf==2.1.0
aliass_free_torch==0.0.6
phaseaug==0.0.2
gradio==3.20.1
